change BootStrap init method to read data (tab delimited text files)
There is still an issue with 3 of the 1746 records (I believe that some
cells in the original spreadsheet have embedded tabs), so I had to
modify the domain class to add/increase the maxSize constraint. I think
I see a way to resolve this, but for now the impact is minor.